// src/typescript/util/vdom/types.ts
export interface AbstractText<T extends TextValue = TextValue> {
  type: NodeKeys.Text,
  value: T,
}
// ...
export type AbstractNode =
  | AbstractText
  | AbstractElement
  | AbstractComponent;

export type AbstractChildren = AbstractNode[] | AbstractNode;

export interface AbstractTag<
  T extends Namespace = Namespace,
  U extends string = string,
> {
  type: NodeKeys.Element,
  namespace: T,
  tag: U,
}

export interface AbstractElement<
  T extends Namespace = Namespace,
  U extends string = string,
> {
  type: NodeKeys.Element,
  namespace: T,
  tag: U,
  properties?: AbstractProperties,
  children?: (AbstractText | AbstractElement | AbstractComponent)[] | AbstractText | AbstractElement | AbstractComponent
}


// Component

export type AnyBareComponent =
  () => AbstractChildren;

export type AnyPropComponent =
  (properties: AnyObject) => AbstractChildren;

export type AnyComponent<T extends AnyFunction = AnyFunction> =
  T extends AnyBareComponent ? T : AnyPropComponent;

export type ComponentProperties<T extends AnyPropComponent> = Parameters<T>[0];

// export interface AbstractFunction<
//   T extends AnyComponent = AnyComponent
// > {
//   type: NodeKeys.Component,
//   component: T,
// }


export interface AbstractComponent<
  T extends AnyComponent = AnyComponent,
> {
  type: NodeKeys.Component,
  component: T,
  properties?: ComponentProperties<T>,
  children?: ComponentProperties<T>['children'],
}